Refactor project into single main module
