--- utils.py
@@ -1,14 +1,0 @@
-import os
-import yaml
-from typing import Any, Dict, List
-
-def load_config(path: str) -> Dict[str, Any]:
-    with open(path, "r", encoding="utf-8") as file:
-        return yaml.safe_load(file)
-
-def discover_csv_files(directory: str) -> List[str]:
-    return [
-        os.path.join(directory, f)
-        for f in os.listdir(directory)
-        if f.endswith(".csv") and os.path.isfile(os.path.join(directory, f))
-    ]
